mercury provides portable atomic counters for us using what ever is available on the current system (e.g. OPA library, stdatomic.h, OSX's OSAtomic.h, the windows API).   while this works fine for C, there are C++ issues when mercury decides to use stdatomic.h. ( see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=60932 ).
for C, both gcc and clang provide an "_Atomic" keyword that stdatomic
uses to declare atomic values:
     _Atomic int foo;

and mercury_atomic.h makes use of that.   unfortunately, for C++
g++ does not define _Atomic and you get compile errors.   clang
does define _Atomic (even though that isn't in the C++ standard).
the gcc people say that you shouldn't use _Atomic in C++ code,
and instead you should use <atomic>, but that isn't part of older
C++ (it is c++11?), so you can't count on it being there.   see
discussion in the 60932 bug report above).

all we really want for shuffler is an atomic counter without having
to recreate all the portability tests that mercury has already done
for us.   to hack around this problem, we provide some C code in
acnt_wrap.* that wraps the mercury API and let our C++ code call that.