rename the seq in rpc in/out to be iseq and oseq to make it easier to tell them apart.
add the rank number of the proc starting an RPC (via HG_Forward) to
the RPC so that the remote side can save it (mainly for debugging).

since we are now passing the iseq and forwarding rank, modify
req_parent_init/req_to_self/req_via_mercury to take an rpcin_t
structure rather than just a seq (we can extract both the iseq
and the forwarding rank from the rpcin_t).

note that rpcin_t is only valid (non-NULL) if we have a handle
from an inbound RPC.   shuffler_send will have both the handle
and rpcin_t pointer set to NULL.