add a delivery thread wakeup threshold value.  we only wake up the delivery thread when the number of items in the delivery queue is larger than the threshold value.   this allows us to batch reqs to the delivery thread and reduce context switching (which is especially costly if we bind each proc to a core -- e.g. mpich "-bind-to core").
a delivery queue flush will wake the delivery thread even if there is
less than the threshold number of requests in the queue (because we want
to flush them all).

note: currently reqs will sit in the deliverq for an unbounded
amount of time waiting for the threshold to be met (or a flush).
we could put a time limit on this if needed (it is just a matter of
changing the delivery thread pthread_cond_wait to a timedwait).

nexus-runner gets a new flag "-h" to set the delivery thread wakeup
threshold value.