add "-j" flag to allow alternate transport for local hops
by default we use na+sm for the local hops, but we can change
this using "-j" to specify a different transport (e.g. "-j ofi+tcp"
which would use TCP over the loopback interface).   this matches
the NEXUS_ALT_LOCAL option from the preload lib (and the code here
is based on that code).

note that we also have the "-x" option (use the remote transport for
local hops too).